feat: category suspense
